lib/util: Add helper for loading eBPF object optionally checking size
The helper can optionally try to estimate the memlock size needed for the
object, and raise the rlimit before doing the load; this makes it possible
to retry with a higher rlimit if the initial load fails.

Also make the program name a parameter to load_xdp_program(), and rename
that to attach_xdp_program().

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>