common.mk: Better dependencies an invocation of lib/ dependencies
For files in lib/, we can just depend directly on the .c and .h files, and
invoke only the needed make commands. This makes the output a bit cleaner,
and also avoids a couple of extra variable expansions.

Also, we don't need the $(MFLAGS) variable for recursive expansion (as that
is only needed for compatibility with old versions of 'make').

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>