Iterate over loaded programs using the pinned program information
This changes the program pinning to pin programs by their program name in a
directory named after the interface. This makes it possible to reliably get
back to the program name for an interface without having to rely on the
program names returned by the kernel.

Also add interfaces in lib/util to iterate over the interfaces that have
programs loaded on them, and use this for getting the feature list, and
printing the xdp-filter status. This also means we no longer iterate over
all interfaces on the machine.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>