configure: Only write config.mk if everything succeeds
Use a temp file to write all the variables to while the checks are running
and only move it into place at the end. This prevents a partial config.mk
from being written if a check fails, which in turn ensures that the
top-level Makefile will re-run the ./configure script on its next
invocation.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>