Makefile: git submodule detect and warn if out-of-sync
It is a reoccuring issue that git submodule gets out-of-sync, which
cause confusion and issues.

 End-user needs to run:  git submodule update

We cannot run this automatically, because there are legimite
reasons for the submodule to be out-of-sync, e.g. when testing
local changes.

Thus, this patch detect SHA-1 doesn't match, and just warns
end-users, together with a hint to call: git submodule update

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>