libxdp: Update function signatures
We want to be able to return an error from the xdp_program__set*()
functions. Specifically, return an error on an attempt to set the run_prio
or chain call actions on an already loaded program.

Also make the xdp_program__btf() getter return type const, to indicate that
the BTF information is not expected to be modified.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>