libxdp: Ensure that internal dispatcher functions always chain call
This adds a special return value for the stub functions defined by the
dispatcher which we make sure is always set in the chain_call_actions for
each function. This is to ensure that the stub functions will always chain
call to the next function, so that if an freplace link is broken, the
subsequent programs will still be reached regardless of the
chain_call_actions set by the program that was removed.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>