Add test scripts and Makefile targets
This is the beginning of a test harness for the various utilities. Each
utility can specify a TEST_FILE in its Makefile which is a shell script
with test definitions (each of which is a shell function).

The test harness will setup a netns-based environment with virtual
interfaces that XDP programs can be loaded onto. Each test function will be
executed and its return value used to determine pass/fail. If the test
passes, output will be supressed, and if it fails it'll be printed after
the test name.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>