lib/common.mk: Support generating man page from README.org
This uses the Emacs man page export support to export the README file in
man page format, then insert date and version information in the final
file. It relies on a few export settings in the README file itself to get
the right file name and a correctly-formatted header. Example:

Add emacs-nox as a build dependency to get the export working.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>