common.mk: only generate man page if Emacs is available
Also remove the emacs package from the Travis config, since it seems the
version in Ubuntu bionic is too old to include a recent enough org-mode to
build the man page. Building the man page as part of CI is not hugely
critical right now, so let's just disable it to unbreak the CI tests.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>