Bump version number and libxdp version
This bumps the libxdp .so-version to 1.0.0, and the tools release to
1.0.0-beta1. We haven't made any releases that have not been marked as
pre-releases (and thus unstable), so it's safe to change the .so version
and all the symbols. Going forward, once we have made the final 1.0.0
release we will obviously need to commit to keeping the API stable,
including with backwards-compatible symbol versioning.

This change also defines the TOOLS_VERSION separately from LIBXDP_VERSION
and moves that definition to its own version.mk to make it possible to
include it from other places. The TOOLS_VERSION will also be the package
version, and we keep a -beta1 suffix to this for now, so we can get a bit
more testing in before tagging the final 1.0.0.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>