xdp-filter: Fix format-string-related compile errors
This fixes a couple of format-string-related compile errors on ppc64 and
s390x architectures. On ppc64, __u64 is defined as unsigned long instead of
unsigned long long, so %llu doesn't work as a format specifier, and we
can't use PRIu64 either, since that also ends up with the wrong definition.
Instead, fix this by explicitly casting to uint64_t, following this commit:

https://github.com/borkmann/lksctp-tools/commit/398f377fbce6cf7c994d525ab0471cb5bc8d832c

The other error seems to be that the compiler is not smart enough to figure
out that the 'flags' element will be filled by bpf_map_lookup_elem() so it
complains about uninitialised variables on s390x. Fix this by initialising
the variable first.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>