libxdp: Skip dispatcher when loading HW mode progs
There is no dispatcher support for offloaded XDP programs, so loading would
always fail in with XDP_MODE_HW. Since the same interface can have both an
offloaded and a non-offloaded XDP program loaded at the same time, there is
not the same need to always enforce a dispatcher when using HW mode. So
instead, allow loading a single program in HW mode, and skip the dispatcher
for this.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>