util/params: Also print libbpf version in --version output
The libbpf version used can have a significant impact on the user
experience, since new BPF features often need to be supported by libbpf
before the BPF object files using them can be loaded. To make it simple for
the user to discover which version of libbpf is in use, let's print the
libbpf version in the --version output.

Unfortunately, libbpf doesn't actually export its own version as part of
its API, so we have to jump through a few hoops to achieve this: If we're
statically linking, we can just embed the version discovered at compile
time (which we parse out from libbpf.map the same way the libbpf Makefile
does), but when linking dynamically the library version can change after
the fact as the library is upgraded. To get an up-to-date version number
without support from the library itself, we parse /proc/self/maps looking
for the full filename of libbpf.so that is mapped into the binary.

Hopefully the need for this can be removed in the future if libbpf starts
exposing a 'libbpf_version()' function directly.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>