xdp-loader: Change to loading programs in native mode by default
Previously we were loading programs without specifying an explicit mode by
default, letting the kernel pick. This causes problems for a couple of
reasons:

- It causes silent fallback from native to generic mode, making it harder
  to predict which mode a program would be loaded in.

- There's an nfp driver bug which makes it harder to unload such programs.

Arguably, the default should have been to just load in native mode from the
start. So let's change the default, but let the user go back to the
kernel-picked mode by specifying '-m unspecified' when loading. Document
the change in defaults in the man page.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>