libxdp: Fix validity checking in xdp_program__detach_multi()
The introduction of HW prog mode made the validity checking in
xdp_program__detach_multi() be slightly inconsistent among the different
modes. This fixes that so that we always check that the program being
detached is actually the one loaded. It also fixes a few fd leaks in the
error paths.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>