xdpdump: Fix signed compare error and logic mistakes in find_target()
The RHEL RPM builder was complaining about mixed sign compares in
find_target(). Turns out there also was what appears to be a copy-paste
mistake in the second parsing of ID numbers, and a mix of strtol() and
strtoul() usage. Fix all of these so we always parse and store program IDs
as unsigned long.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>