xsk: export inline helpers as symbols in libxdp
This helps people writing language bindings to not have to rewrite C
wrappers for inline functions in the headers. We force inline the
definition from the header for C and C++ consumers, but also export a
symbol in the library for others. This keeps the performance advantages.

Also see
https://lore.kernel.org/bpf/CAJ8uoz0QqR97qEYYK=VVCE9A=V=k2tKnH6wNM48jeak2RAmL0A@mail.gmail.com/
for some context.

extern inline is used as it's slightly better since it warns when an
inline definition is missing.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>