xdp: Prevent double close of dispatcher prog fd
The dispatcher prog is opened using the open_file path, then loaded onto
the interface, where down the call chain xdp_program__load is called.
This function directly invokes the xdp_program__fill_from_fd helper, but
fails to take into account that it consumes the fd that is passed in (as
it is stashed in the xdp_prog structure).

In all other call sites (wrapped using xdp_program__from_fd), the fd
being consumed is acquired right before it is passed in, so the
ownership transfer works as expected.

All in all, this leads to release paths of the xdp_prog and the bpf_prog
both trying to call close on the same fd, where the xdp_program__close
wins due to it being early in the call chain, and the
bpf_object__unload -> bpf_program__unload's close call gets -EBADF.

Fix this by duplicating the descriptor before passing to the helper.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>