xdp: add async_close helper
This is required for the case where closing a fd corresponding to a
bpf_link can block the program. We try a best effort async_close,
falling back to the blocking close when we get an error.

We also perform feature detection once during the initialization of the
library (either during program startup or when the .so is loaded), to
avoid doing the mmap dance each time and failing.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>