libxdp: Fix getting map ID for program when struct bpf_prog_info changes
If the size of struct bpf_prog_info changes, libbpf may have been compiled
with a different notion of what that size is supposed to be compared to
what we have in our own header file. This means we can't just use the
'data' member of struct bpf_prog_info_linear to get that information.
Thankfully, libbpf helpfully provides us with its notion of the length of
that struct, and thus the offset of the data member. So use that to compute
the offset instead.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>