Include text of all licenses
We license different parts of xdp-tools under different licenses: The tools
themselves are GPL-2.0, while libxdp is dual-licensed LGPL-2.1 and
BSD-2-Clause. However, only the text of GPL-2.0 was included. Fix this by
including all three licenses, importing them all from the kernel tree (with
the SPDX header instructions at the top of the files).

The existing GPL text was identical to the one in the kernel tree except
for whitespace changes and the fact that it expands LGPL into "Lesser
General Public License" instead of "Library General Public License". Since
it's LGPL-2.1 we use (which is the version that introduced the renaming),
use the version of the GPL-2.0 text that we already had in the repository,
but keep the instructions in the header from the version in the kernel.

Reported-by: Magnus Karlsson <magnus.karlsson@intel.com>
Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>