configure: rework LIBBPF_DIR parameter
Rework the way the LIBBPF_DIR parameter is used: rather trying to
dynamically link against the libbpf version in LIBBPF_DIR, which most
likely never worked anyway, use it as a way of specifying a separate
location for the libbpf sources that will be statically linked against in
the build system. This is particularly useful when embedding libxdp in a
different repository that also uses libbpf as a submodule, so that libxdp
will be linked against the same libbpf version as the rest of the code in
the embedding repository.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>