libxdp/Makefile: Add XDP_OBJS as prerequisite of library objects
This makes sure the libraries are always rebuilt after the xdp objects
change, which avoid a rebuild cycle where the generated xdp-dispatcher is
newer than libxdp.a. This mostly impacts embedding of libxdp, but shouldn't
hurt for regular builds, and if we ever end up depending on the XDP objects
during build this change will be needed anyway.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>