libxdp: document kernel and BTF version requirements
The multiprog support relies on specific kernel features introduced in
kernel versions 5.6 and 5.10. However, this was not actually documented
anywhere, leading to user confusion. Fix this by adding a section to the
man page with the commits and kernel versions needed for both multiprog and
AF_XDP. Also note that BTF is needed for the BPF programs to work with
multiprog.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>