libxdp: Load embedded bundled object files when compiled as static library
When loading the bundled XDP objects, we look up the files by name in the
search directory paths. However, for the statically linked version of the
library we can't be sure that the object files are shipped with the static
binary, so we embed a version of the object file into the .rodata of the
static library as well. This commit adds code to load the dispatcher and
XSK object file from these embedded sections if libxdp is compiled as a
static library.

We prefer en embedded version to make sure multiple static copies of libxdp
don't interfere with each other, nor with a system-installed dynamic
library.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>