configure: generalise feature probing and add new function tests
Upstream libbpf has deprecated a lot of functions we use, so we need to add
a lot of new tests for function existence. To avoid having to add a lot of
boilerplate, generalise the feature probing function in configure to be
parameterised by function name, and update the Makefile to dynamically
generate the defines for the feature list (using the existing HAVE_FEATURES
variable).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>