libxdp: fix program lookup for legacy
If the legacy loading prodedures are used, xsk_lookup_program will
fail even if there is a program loaded on the interface because
xdp_multiprog__next_prog fails. To rectify this introduce a check for
legacy mode which if true, will skip the xdp_multiprog__next_prog call
and ensure the legacy program is returned.

Fixes: ea481a18ed07 ("libxdp: move AF_XDP functionality from libbpf")

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>