libxdp/xsk: Add compatibility check for bpf_map_create()
The new low-level APIs for creating maps and programs, bpf_map_create() and
bpf_prog_load() were added at the same time, so add a compatibility check
for bpf_map_create() and use both if that succeeds.

We can't just redefine both functions if the compatibility check fails,
because there was an old bpf_prog_load() function in older versions of BPF.
So we define our own wrapper function for creating the program and have
that call the new bpf_prog_load() version if it exists.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>