libxdp: Allow falling back to single-program attachment for loaded programs
When the libxdp compatibility checks for the dispatcher fails, we can fall
back to just attaching a single program. However, the
xdp_program__attach_single() function assumes that the program had not been
loaded into the kernel yet, which is not always the case. In the fallback
case there's not really any reason to not just try anyway, so just skip the
program load if we already have an fd and try a regular program attach.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>