xdp-trafficgen: Add TCP flow suport
Idea: Open the TCP connection from userspace, but install an XDP program
that will intercept the ACK and capture the window parameters. Then, once
the flow has been established, run an XDP program with prog_test_run that
generates data packets as fast as it can, up to the rwind of the receiver.
The XDP program on the interface is kept in place to intercept ACKs and
update the rwin + ACK seqno.

Status: The basic idea works - connecting, and taking over the transmission
from XDP, including handling ACKs. However, after a while the receiver
stops acking packets, and the sender gets stuck at the end of the RWIN.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>