libxdp: unmap rings when umem deleted
xsk_umem__create() does mmap for fill/comp rings, but
xsk_umem__delete() doesn't do the unmap. This works fine for regular
cases, because xsk_socket__delete() does unmap for the rings. But for
the case that xsk_socket__create_shared() fails, umem rings are not
unmapped.

fill_save/comp_save are checked to determine if rings have already be
unmapped by xsk. If fill_save and comp_save are NULL, it means that
the rings have already been used by xsk. Then they are supposed to be
unmapped by xsk_socket__delete(). Otherwise, xsk_umem__delete() does
the unmap.

Note that when I activate Linux style in Emacs, it finds one place in
a comment with an extra space that it fixes as a bonus.

Fixes: ea481a18ed07 ("libxdp: move AF_XDP functionality from libbpf")
Signed-off-by: Cheng Li <lic121@chinatelecom.cn>