check for zlib before libelf in configure
The libelf library requires zlib. If zlib isn't installed the pkg-config
check for libelf fails silently even if libelf is install leading to a
lot of confusion.

This patch moves the zlib check before libelf to catch this case.

Example of the error case where this was occuring for us:

```
% pkg-config --print-errors --exists libelf
Package zlib was not found in the pkg-config search path.
Perhaps you should add the directory containing `zlib.pc'
to the PKG_CONFIG_PATH environment variable
Package 'zlib', required by 'libelf', not found
```

Signed-off-by: Marcus Barczak <mbarczak@gmail.com>