configure: Support LIBBPF_UNBUILT envvar for submodule embedding
Add support for the LIBBPF_UNBUILT environment variable in the configure
script. When this is set, feature detection will assume that the libbpf
library in $LIBBPF_DIR has not yet been built (and so can't be linked
against). Instead, the feature detection will copy over the include files
from $LIBBPF_DIR/src into a synthetic include dir, and run feature test
compiles against those while instructing the linker to ignore undefined
symbols. This way, the header files can be probed for the right functions
without having to link anything.

This feature is meant to be used when embedding xdp-tools as a submodule in
another git repository; in this case, LIBBPF_DIR can be set to make sure
that libxdp is linked against the same libbpf instance as the surrounding
code, in which case LIBBPF_UNBUILT ensures configure won't give the wrong
results.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>