configure: Don't check for bpf_set_link_xdp_fd_opts()
The configure script checks for the presence of the old_fd argument in
bpf_set_link_xdp_fd_opts() to determine if it can use the system libbpf.
However, this function has been deprecated in newer versions of libbpf,
which means we refuse to use system libbpf versions 0.8 and above.
Fortunately, we can just remove this check, as we also check for
bpf_map__set_initial_value() which was added after the old_fd support.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>