xsk: Add possibility to run AF_XDP with multiprog
Add a libxdp flag XSK_LIBXDP_FLAGS__USE_MULTIPROG
that allows user to attach xsk XDP program to a multiprog
(or by the means of netlink, if not supported)
instead of using bpf_link.

Also add APIs that allow convenient detachment of an XDP program
associated with an AF_XDP socket.

Signed-off-by: Larysa Zaremba <larysa.zaremba@intel.com>