configure: Expand bracket combination when copying libbpf headers
The Dash shell (which is used for /bin/sh on Debian and derivatives)
doesn't support the bracket-based header file expanding pattern we used to
copy over libbpf headers when LIBBPF_UNBUILT is set. Fix that by manually
expanding the brackets, and also complete the incomplete comment above the
code doing the copy.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>