headers: Update kernel btf.h header file
The libbpf btf.h header file shipped with libbpf version 1.0.0 fails to
build with older versions of the btf.h header file from the kernel because
of a missing struct btf_enum64 definition. This was not added in the
upstream kernel until the 6.0 cycle, so this header update is not yet in
any released version of the kernel. Technically this makes it a bit
premature to import this, but seeing as libbpf already introduced a
dependency on the header file, let's do so anyway to work around the buggy
upstream.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>