libxdp: Add support for XDP frags
To enable the use of XDP with large MTUs, the kernel added support for
addressing multiple packet fragments from XDP, AKA "XDP frags" in kernel
version 5.18. This support is opt-in from the program side since it changes
the semantics of the XDP program: The full packet data is no longer
guaranteed to be addressable through the data and data_end pointers of the
XDP context object, and the XDP program has to use a new API to access
packet data after the first fragment.

An XDP program opts in to using this support by setting the
BPF_F_XDP_HAS_FRAGS flag when loading a program, and the kernel only allows
attaching an XDP program to a large-MTU device if this flag is set.
However, because libxdp did not understand the flag, frags support could
not be used at all with libxdp (even if a program set the flag, this would
not be propagated to the dispatcher, so from the kernel PoV it was never
set).

To fix this, we need to make libxdp aware of the frags support, and
properly propagate the flag. Since libxdp, by its nature, allows attaching
multiple programs, some care has to be taken when mixing them: the frags
flag should only be propagated to the dispatcher if all programs attached
to it set the flag. Otherwise, a non-frags aware program may be exposed to
a multi-frag packet and malfunction in a subtle way. The converse, however,
is OK: A frags-aware program can be loaded without the frags flag set, the
resulting dispatcher just won't be attachable to a device with a big MTU.

Because it is not possible to query the kernel and learn whether an
already-loaded program has the frags flag set, we need to store this
information ourselves. This means we have to change the format of the
dispatcher data structure, so this patch also bumps dispatcher version
field. Unfortunately, this means that this patch breaks
backwards-compatibility with older versions of libxdp: if a program is
loaded with libxdp after this patch, older versions of libxdp will refuse
to touch that dispatcher. Conversely, if this version of libxdp encounters
an older dispatcher, it will load it and display the component programs,
but it will refuse to  upgrade it to the newer version without an explicit
unload first.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>