libxdp: mount bpffs if it is not found
The dispatcher functionality in libxdp relies on bpffs being available to
pin the links and programs attached to the dispatcher. Which means that if
no bpffs is mounted, any libxdp operation is going to fail. To prevent
this, try mounting a new bpffs if it is not already mounted.

The 'mount --make-private' and self-binding of the mount point logic is
shamelessly lifted from iproute2 (but modified to better fit with the
coding style of libxdp).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>