util: Make bailing out when there's no bpffs optional
Not all of the command-line utilities actually need a bpffs to be mounted
to work. In fact, only xdp-filter absolutely relies on it, xdp-loader can
defer to the libxdp handling of bpffs instead (which we're fixing
separately). So refactor the util code so each utility can decide whether
it wants to be run with a mounted bpffs or not.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>