libxdp: Only auto-mount bpffs if LIBXDP_BPFFS_AUTOMOUNT is set
If there is not bpffs mounted, automatically mounting one may end up
leading to other hard-to-debug errors. In particular, when using network
namespaces with 'ip netns exec', the mount point will go away as soon as
the caller exits, which will make the programs attached to the dispatcher
disappear.

For this reason, it's better to be conservative when no bpffs is mounted.
So instead of always auto-mounting, we check the environment for the
LIBXDP_BPFFS_AUTOMOUNT envvar. If this is set to '1', we attempt the
automount. If not, we instead fail the compatibility check for the
dispatcher, which means we'll fall back to attaching without the dispatcher
if we're only attaching a single program, and bail out otherwise.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>