libxdp: Make sure to set the the program autoload when loading a program
When loading an xdp_program, we just load the underlying BPF object;
however, if that has previously had autoload disabled on the program we're
trying to load, the program won't actually be loaded. Since the load
corresponds to a particular program this is surprising behaviour, so make
sure we turn on program autoload for the program we're working with before
loading the BPF object.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>