xdp-bench: Skip cpumap and devmap prog tests on old kernel
The xdp-bench utility 'redirect-cpu' action relies on the ability to attach
programs to a cpumap. The other redirect actions implement compatibility
fallbacks, but the cpumap program does not. So for now, just skip the the
cpumap tests on older kernels. Also skip the egress tests that require
devmap program attachment for the redirect-map and redirect-multi tests.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>