configure: Try to auto-detect versioned clang binaries
The clang and llc binaries can be installed as versioned binaries named as
clang-$VERSION. In particular this happens when using the Debian and Ubuntu
packages from the apt.llvm.org repository. To make installation a bit
friendlier, teach the configure script to autodetect such binaries, by
trying each one in order from the newest (clang-16, llc-16) to the
oldest (clang-10, llc-10). We only do this if no explicit path is supplied
in the environment, and only if the non-versioned binary doesn't exist.
I.e., the objective here is to only do autodetection in cases where we
would otherwise fail configure because no binary was found.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>