libxdp: Support propagating BPF_F_XDP_DEV_BOUND_ONLY flag to dispatcher
Add support to libxdp for propagating the BPF_F_XDP_DEV_BOUND_ONLY flag
from component programs to the dispatcher. We just set the flag on the
dispatcher if any program has it set. This works even for non-devbound
programs, but has the side effect of binding all component programs to that
interface (preventing them from being re-attached to a different device
later even if they're not devbound).

To avoid this extra binding we'd need to first attach programs to a
separate dispatcher that's not devbound, then re-attach them to the
dev-bound dispatcher. We'll leave this for later, though, as the initial
thing is good enough for testing.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>