xdp-bench: Support reading HW metadata in 'drop' and 'tx' actions
Add a new --hw-meta flag that will cause the XDP-bench program to read HW
metadata (using the kfuncs available for this) before transmitting or
dropping the packet. When enabled, the 'drop' counter will be incremented
only if reading the RX-hash succeeds, as a way for userspace to check that
it's working.

We don't perform any feature detection, so this patch causes the XDP
program to fail to load on kernels that don't support the metadata
kfuncs (even if --hw-meta is not set).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>