xdp-bench: Use basic_program_mode enum in BPF code as well
Instead of converting the basic_program_mode enum into boolean variables,
just use it directly in the BPF code. This makes it easier to add new modes
in the future, as the boolean variables quickly explodes. Move the struct
definition itself into its own shared header file so it's kept in sync.

This includes a small functional change in that we no longer perform the
"touch data" action when we're swapping MACs. However, since swapping MACs
touches the same part of the packet data, the performance impact of this
should be negligible (it's just a single inline branch that's being
removed).

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>