libxdp: Fix error path in xsk_update_prog_refcnt()
Coverity pointed out that we leak a memory allocation if taking the lock
fails in xsk_update_prog_refcnt(). Fix this by freeing in the error path.
Also switch the allocation to calloc() instead of malloc() to silence an
"unitialised variable" warning; the warning is technically a false positive
caused by Coverity not understanding the semantics of
bpf_map_lookup_elem(), but we use calloc() elsewhere anyway, so might as
well switch to it to make sure we zero-init the memory in any case.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>