.github/workflows: Set DID_UNSHARE for old kernel versions
It seems that running the libxdp selftests in a separate namespace with
'unshare' breaks on older kernels. So export the DID_UNSHARE variable as
part of the test environment for those older kernels, thus making the test
script skip the call to 'unshare'.

This doesn't actually work unless we also propagate the environment vars
into the test runner script running in the virtual machine, so add a
generic mechanism to do this, and propagate all the environment vars set in
the selftests.yml file.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>