xdp-trafficgen: Port to using BPF skeletons
In preparation for plugging in to the xdp_sample util functions, port
xdp-trafficgen to use BPF skeletons to load its programs. This requires
moving the trafficgen state to an explicit state map instead of using a
global variable, as using skel->bss doesn't work after load when the
programs are loaded with libxdp.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>