libxdp: Factor out kernel compat check function and expose in libxdp_internal.h
Factor out the libxdp compatibility check function so it can be called
standalone (without a multiprog argument), and expose it in
libxdp_internal.h so we can call it from the libxdp selftests. Add a binary
to the selftests that just calls this function and expose that as a skip
function in the test runner, so we can skip subtests that don't make sense
without the dispatcher.

Previously we cached the result in the xdp_multiprog struct, move that to a
static variable so we only run the compatibility check once per instance.
Check for bpffs availability on every invocation, though, as this can
change between each invocation.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>