libxdp: Make sure to reset new_mp pointer on detach retry
Coverity pointed out that when we hit the retry path in
xdp_program__detach_multi() we may start out with a new_mp pointer that is
set to an old value that was freed, but the pointer has not been reset. Fix
this by moving the NULL-initialisation of the pointer to after the retry
label.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>