configure: Fix function detection for btf__type_cnt()
The configure check for btf__type_cnt() swapped two arguments in the
check_libbpf_function() call in the configure script, leading to the check
failing if LIBBPF_CFLAGS is non-empty. Make sure the arguments are in the
right order, and also pass a proper NULL parameter in the check instead of
a 0.

Reported-by: Holger Hoffstätte <holger@applied-asynchrony.com>
Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>