util: Move program locking out of dispatch_commands()
The dispatch_commands() helper takes a program lock and holds it while the
command is being executed. This was added for the use of xdp_filter, and is
really only needed for that; none of the other utilities really need
locking to synchronise with itself (and libxdp has its own locking for the
dispatcher).

Move the locking out of dispatch_commands() and instead add explicit lock
and unlock calls to each of the xdp-filter commands. While doing so, also
move the locking to use the same directory-fd based approach as the libxdp
locking uses. This has the benefit of not writing any lock files, so the
lock is automatically cleared when the application exits, even if it
crashes. This should fix any issues with stale lock files sticking around,
which we did not have any good way to clean up.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>