util/compat.h: Fix compilation errors with unknown functions
One of the compatibility functions in compat.h was not declared as static
inline, which causes the compiler to complain if it is not used. Fix this
and also add #include directives to the file so it works even when included
by another file that doesn't include all the libbpf headers.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>