xdp-bench: restucture the program options
Instead of having the program operating modes of the basic runner be part
of the BPF program (and selected with vars in the rodata), turn each mode
into its own program. This way the selection can be done in userspace by
just selecting the right program, and we reduce the complexity of the
branches on the BPF side, and rely less on dead code elimination by the
verifier.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>