filterc: Convert missing store instructions
The store instructions need to be guarded, i.e., checked that they only
access data within the packet, because we need to pass the verifier.

Linux supports legacy packet access instructions that were carried over
from cBPF. However, they are only supported for programs where the context
is an skb, not for XDP programs.  Therefore, rewrite legacy packet access
instructions to a series of eBPF instructions.

Document register usage for the future. We prefer caller-saved registers at
the moment because they don't need to be saved, i.e., pushed to the stack,
in the JITed function, and we don't call into other functions anyways.

Signed-off-by: Felix Maurer <fmaurer@redhat.com>