filterc: Optimize jumps
When one of the branches goes to the next insn, it is not necessary to emit
two eBPF insns. We can either leave out the second jump if the false case
is the next insn, or we can invert the condition and leave out the second
jump if the true case is the next insn.

Signed-off-by: Felix Maurer <fmaurer@redhat.com>