.github/workflows: Run mkarchive.sh directly from release workflow
We don't install all the prerequisites for building to the release workflow, and
because the 'make archive' target now inadvertently depends on config.mk, that
ends up trying to run configure, and then fail. Avoid this by just running the
mkarchive.sh script directly, now that it supports being run without a version
argument.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>