libxdp: Fix xdp prog memory leak in xsk_setup_xdp_prog
In the xsk_setup_xdp_prog function, the xsk structure
temporarily takes ownership of an xdp_prog and stores it in
ctx->xdp_prog. However, the allocated memory is not freed
in xsk_destroy_xsk_struct, leading to a memory leak. This
commit addresses the issue by adding a call to
xdp_program_close to properly release the allocated
xdp_prog memory.

Signed-off-by: Ric Li <ming3.li@intel.com>