lib/testing: Simplify test_runner.sh script
The test_runner.sh script is derived from the test-netns script originally
developed for the xdp-tutorial, and still contains a lot of the legacy from
there. Specifically, it uses a state directory with file system-based state to
support interactive execution, saving state across multiple invocations of the
script.

However, the way we execute the tests here, we always clean up everything in the
same run, so all this state saving complexity is unnecessary. Get rid of it and
store most things in global variables, so we save this complexity. This also
makes it possible to support creating multiple namespaces at the same time,
which will be needed for adding tests for the xdp-forward tool. For now, we just
create two namespaces (with matching veth pairs), but only use one of them, but
future tests will use both for forwarding tests.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>