testing/test_runner: Only set errexit during setup
We rely on `set -o errexit` to abort tests if any of the setup commands fail.
However, we don't want this turned on during test execution, since we rely on
explicit checks using check_run() to check the output of certain commands in the
tests. So move the errexit to cover only the setup function.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>