lib/common.mk: support eBPF programs in nested directories
The current build system sets the name of the generated skeleton directly
to the file's name minus the .skel.h suffix. In case the eBPF program
was in a nested directory, this leads to syntax errors in the skeleton
file. This commit support ebpf programs in nested directory by poping out
the directory part in the eBPF program name without breaking the build
of currently available eBPF program.

Signed-off-by: Jalal Mostafa <jalal.a.mostapha@gmail.com>