util/xdp_sample: Count total packets correctly
The xdp_sample library functions were counting the total number of
packets by adding the pps number regardless of the sample interval. If
the interval is larger than one second, this leads to a miscount. Add a
collection of functions that count packets instead of pps and use those
for totals. Using an extra set of (trivial) helpers this way makes the
call sites nicely symmetrical between pps and packet calculations.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>