util/xdp_sample: Print PPS values in summary line instead of totals
The Summary line output was printing the total values instead of the PPS
values, which led to wrong values being printed now that we account the
total values correctly. Fix the output to print the PPS values instead
of the totals.

Add a new err_pps variable to stats_output.totals to record accumulated
errors from the various count sources.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>