libxdp: Fix incorrect setsockopt xdp_umem_reg size.
The commit in the fixes tag made an uneccesary and incorrect
attempt at backwards compatibility with kernels that do not
support tx_metadata_len.

The kernel will either expect a size matching its version of
struct xdp_umem_reg, if it encounters a shorter size, clamp it to
24 byte struct xdp_umem_reg_v1 [0].

The code in question attempted to provide a 28 byte length when
tx_metadata_len was not set, which would invoke the kernel v1
processing.

A side effect of v1 processing is that flags are no longer
processed by the kernel, which would cause a regression for
applications that rely on them.

This causes issues downstream as revealed in the Open vSwitch
system testsuite. I have confirmed that by removing this
conditional, the test passes both on 6.7 and newer kernels.

0: https://github.com/torvalds/linux/blob/0ad2507d5d93/net/xdp/xsk.c#L1369-L1372
Fixes: 44b5e00692db ("libxdp: Add tx_metadata_len/xsk_umem__create_opts() to support AF_XDP Tx metadata")
Reported-at: https://launchpad.net/bugs/2098005
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>