xdp-dump/xdpdump.h: Increase MAX_CPUS to 512
Previously, MAX_CPUS was set to 256. This causes the tool to exit with a failure on systems with more than 256 CPUs.
For example, an AMD EPYC server with two AMD EPYC 9684X 96-Core Processors reports 384 CPUs to the kernel due to hyperthreading.
This fix allows such systems to run the xdpdump utility.

Signed-off-by: Lucas Crijns <lucascrijns@gmail.com>