xdp-trafficgen: Load a dummy XDP program if needed
Add logic in xdp-trafficgen to load a dummy XDP program on devices that
need this to be able to send packets via ndo_xdp_xmit(). We use a
hard-coded list of driver names to determine if an interface needs such
an XDP program, and check the XDP feature flags to skip the loading if
the NDO_XMIT feature is already set.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>