xdp-trafficgen: Error out if interface doesn't support sending packets
If a device doesn't support sending packets via XDP, xdp-trafficgen will
still happily load and try to send packets, but no packets will actually
be sent. Add a check for the NDO_XMIT XDP feature flag in
xdp-trafficgen, and error out if the flag it not set. This should make
it clearer to users why things don't work.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>