libxdp: add device binding support
Kernel 6.3 supported for some NIC offloads for XDP programs.
The feature in XDP is known to be XDP hints. XDP hints are only
supported if the XDP program is bound to the NIC device using the
BPF_F_XDP_DEV_BOUND_ONLY binding flag.

Similarly to BPF_F_XDP_HAS_FRAGS, the device binding flag is
only possible for the dispatcher if all programs has the flag.
Otherwise, it should fail.

The flag is recored using a new flag in `struct xdp_program`
and `struct xdp_multiprogram`.

Signed-off-by: Jalal Mostafa <jalal.a.mostapha@gmail.com>