xdp-forward: added support for 802.1Q VLANs
This commit adds support for VLANs. All the scenarios
of packets forwarding from/to interfaces are supported:

* untagged inf -> untagged inf
* tagged inf -> tagged inf
* untagged inf -> tagged inf
* tagged inf -> untagged inf

Unfortunately, this adds roughly 4% performance overhead
for all the scenarios.

Since patch for kernel this is based on, hasn't been
merged yet this requires patching kernel manually. When
running this version of xdp-forward on unpached kernel
all the route lookups (`bpf_fib_lookup`) will return
`-EINVAL` and so, will be passed to regular kernel stack.