xdpdump: Adapt the behavior of util.c find_bpf_file() to the libxdp.c variant
In contrast to the libxdp.c variant, the function find_bpf_file() in the util.c file does not use the environment variable LIBXDP_OBJECT_PATH to overwrite the default path, but always uses the default variable BPF_OBJECT_PATH, which is set at compile time.
To unify the behavior, the behavior of the util.c variant is adapted to that of the libxdp.c variant.

Signed-off-by: Lukas Schreckenberg <lukas.schreckenberg@iosb-ina.fraunhofer.de>