libxdp: Prevent blocking of programs that modify the packet pointer
The verifier grew a check for global subprograms that invalidate the
packet pointer:

https://lore.kernel.org/all/20241210041100.1898468-1-eddyz87@gmail.com/

This affects libxdp, because the stub programs that are replaced by the
user programs do not modify the packet pointer. Implement the workaround
suggested in the posting above, by adding a no-op call to
bpf_xdp_adjust_tail() to the stub programs being replaced. This does not
affect performance, nor backwards compatibility of the dispatcher, since
the stub programs are never actually executed when libxdp is used. The
only effect will be that on newer kernels, a version of libxdp that
includes these changes will work with programs that modify the packet
pointer, while older versions of libxdp won't.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>