xdp-trafficgen: Fix automatic loading of xdp_pass program
The automatic loading of an xdp_pass program in xdp-trafficgen would
fail on systems that don't allow multi-attach of programs, because it's
attaching an already-loaded program. Switch to loading a separate
program from the dispatcher BPF object file. Also check if an XDP
program is already loaded on an interface, and don't try to load a new
one if it is.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>