lib/util: Import xdpsock code from bpf_examples
This includes the xdpsock code from the bpf_examples repository into
lib/util. The code is included as-is, with only a couple of small
changes to make the code compile (fixing "mixed-sign compare" compiler
warnings). The following commits will refactor the code into a library
to be re-used by xdp-bench and xdp-trafficgen.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>