test_runner: Wait for background output file to appear before moving it
On slow test execution systems (vm-in-vm setups), spawning a background
process can take so long that the output file is not created before the
test script tries to move it into place. This in turns leads to lost
output and subsequent test failures.

Fix this by waiting on the temporary file to appear before moving it.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>