xdp-loader: Make feature flags printing future-proof
In netdev.h, NETDEV_XDP_ACT_MASK is declared as private, so it should
only be used inside the kernel, and according to Jakub Kicinski then
"user space should not care about it"[1].

  linux$ git grep -B1 NETDEV_XDP_ACT_MASK include/
  include/uapi/linux/netdev.h-    /* private: */
  include/uapi/linux/netdev.h:    NETDEV_XDP_ACT_MASK = 127,

In this specific case, then this code is buggy. If a new feature flag
is added to the uapi header, and hence included in NETDEV_XDP_ACT_MASK,
and xdp-tools is not updated, or an older version is re-build in a
distro, with newer linux headers, then NETDEV_XDP_ACT_MASK can contain
a new feature flag, and if the new flag is observed, then it will neither
be checked nor trigger the unknown feature printing.

This patch moves the flags into a constant array, and while looping over
them in iface_print_xdp_features(), it builds a local mask consisting only
of the tested flags. The local mask is used to replace NETDEV_XDP_ACT_MASK.

[1] https://lore.kernel.org/netdev/20230614211715.01940bbd@kernel.org/

Signed-off-by: Asbjørn Sloth Tønnesen <ast@fiberby.net>