configure: Avoid populating .rodata in libbpf test
When checking for a working libbpf, the test includes constant strings
which end up in .rodata. This can cause compilation failures when
LDFLAGS contain values that enforces position-independent code. We don't
really need the string values, so just replace the constants with NULL
values in the test program to avoid the error.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>